+++ src/nespresso/bot/handlers/admin/commands/logs.py
@@ -6,7 +6,7 @@ from nespresso.bot.lib.messaging.stream import (
     ReceiveMessage,
     SendDocument,
 )
-from nespresso.core.configs.paths import PATH_LOGS
+from nespresso.core.configs.paths import PATH_BOT_LOGS
 
 router = Router()
 
@@ -15,4 +15,6 @@ router = Router()
 async def CommandLogs(message: types.Message) -> None:
     await ReceiveMessage(message)
 
-    await SendDocument(chat_id=message.chat.id, document=types.FSInputFile(PATH_LOGS))
+    await SendDocument(
+        chat_id=message.chat.id, document=types.FSInputFile(PATH_BOT_LOGS)
+    )