+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.1.0] - 2025-05-20 - Vlad
 ### Added
 - API for receiving NesUser data
+- Dependense of API and Bot containers on DB
 
 ### Updated
 - Tg messages translated to eng