+++ CHANGELOG.md
@@ -22,6 +22,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ---
 
 ## [0.1.0] - 2025-05-20 - Vlad
+### Updated
+- Separate logging for bot logic, aiogram, API. Bot's console logging includes errors from aiogram
+
+## [0.1.0] - 2025-05-20 - Vlad
 ### Added
 - API for receiving NesUser data
 - Dependense of API and Bot containers on DB